* dev_OpenMP:
  :green_heart: updating the CI test to run for a merge_request
  :bug: the number of block has to stay as an int
  :bug: explicit conversion to size_t for the block size
  :bug: avoiding a spurious multiplication by nf and switching to size_t for the block_size
  :racing_car: adding submission scripts for zenobe weak and strong

+++ src/SwitchTopo.hpp
@@ -103,17 +103,17 @@ class FLUPS::SwitchTopo {
     void setup_buffers(opt_double_ptr* _sendBuf,opt_double_ptr* _recvBuf);
     void execute(opt_double_ptr v, const int sign);
 
-    inline int get_BlockSize () const {
+    inline size_t get_BlockSize () const {
         // the nf at the moment of the switchTopo is ALWAYS the one from the output topo!!
         const int nf = _topo_out->nf();
         // get the max block size
-        int maxsize = 0;
+        size_t maxsize = 0;
         for (int ib = 0; ib < _inBlock[0] * _inBlock[1] * _inBlock[2]; ib++) {
-            maxsize = std::max(maxsize,_iBlockSize[0][ib]*_iBlockSize[1][ib]*_iBlockSize[2][ib])* nf;
+            maxsize = std::max(maxsize,((size_t)_iBlockSize[0][ib])*((size_t)_iBlockSize[1][ib])*((size_t)_iBlockSize[2][ib])* ((size_t)nf));
         }
         // get the max block size
         for (int ib = 0; ib < _onBlock[0] * _onBlock[1] * _onBlock[2]; ib++) {
-            maxsize = std::max(maxsize,_oBlockSize[0][ib]*_oBlockSize[1][ib]*_oBlockSize[2][ib])* nf;
+            maxsize = std::max(maxsize,((size_t)_oBlockSize[0][ib])*((size_t)_oBlockSize[1][ib])*((size_t)_oBlockSize[2][ib])*((size_t)nf));
         }
         // return
         return maxsize;
